Chai Handler Stub almost complete, Bye Bye auth.py
